YOLO

+++ frontend/tethys/src/pages/alerts.tsx
@@ -1,23 +1,25 @@
-import { Title, Center, Container, Text, Stack, Group, Button, Divider, Table, ActionIcon, MultiSelect } from "@mantine/core";
+import { Title, Center, Container, Text, Stack, Group, Button, Divider, Table, ActionIcon, MultiSelect, Code, Space } from "@mantine/core";
 import { DatePicker } from "@mantine/dates";
 import { IconX } from "@tabler/icons-react";
 import { useEffect, useState } from "react";
-import type { GeoJSON as GeoJSONType } from "leaflet";
 import { supabase_s } from "./_app";
 import { useAtom } from "jotai";
-import { authenticated_s, currentUser_s } from "./state";
+import { authenticated_s, currentUser_s, userAlerts_s } from "../state";
 var _ = require('lodash');
+import { fetchAlertDatesByUser, fetchRiversData } from "@/api/supabase";
+import { River } from "./riverMap";
+
+export interface AlertDateRange {
+    startDate: Date
+    endDate: Date
+    river: string
+}
 
 export default function Alerts() {
 
 
     const dateOptions = { month: '2-digit', day: '2-digit', year: 'numeric' };
 
-    interface AlertDateRange {
-        startDate: Date
-        endDate: Date
-        river: string
-    }
 
 
     const [riverList, setRiverList] = useState<string[]>([])
@@ -28,63 +30,46 @@ export default function Alerts() {
     const [currentUser] = useAtom(currentUser_s)
 
     const [dateRange, setDateRange] = useState<[Date | null, Date | null]>([null, null])
-    const [alertDateRanges, setAlertDateRanges] = useState<AlertDateRange[]>([])
+    const [alertDateRanges, setAlertDateRanges] = useAtom(userAlerts_s)
 
     useEffect(() => {
-        const fetchMapData = async () => {
-            await supabase.from('rivers').select().then((data) => {
-                if (data.data) {
-                    setRiverList(data.data.map((item) => (item.name)))
-                }
+        const fetchData = async () => {
+            await fetchRiversData(supabase).then((rivers: River[]) => {
+                setRiverList(rivers.map((item) => (item.name)))
             })
-        }
-        fetchMapData()
-
-        const getAlertDates = async () => {
-            const { data, error } = await supabase
-                .from('alerts')
-                .select('*')
-                .eq('user_id', currentUser?.id)
-
-            if (error) {
-                console.log(error)
-            } else {
-                const alertDateRanges = data.flatMap((r) => ({
-                    startDate: r.start_date,
-                    endDate: r.end_date,
-                    river: r.river,
-                }))
-                setAlertDateRanges(alertDateRanges)
+            if (currentUser !== null) {
+                await fetchAlertDatesByUser(supabase, currentUser?.id).then((alertDateRanges: AlertDateRange[]) => {
+                    setAlertDateRanges(alertDateRanges)
+                })
             }
         }
-        getAlertDates()
+        fetchData();
 
     }, [authenticated])
 
 
-
-    const addRow = async (dateRange: [Date, Date]) => {
+    const addRow = async (dateRange: [Date | null, Date | null]) => {
         if (authenticated) {
 
-            const selectedAlerts = selectedRivers.flatMap((r) => ({
-                startDate: dateRange[0].toLocaleDateString('en-US', dateOptions),
-                endDate: dateRange[1].toLocaleDateString('en-US', dateOptions),
-                river: r
-            }));
-            console.log(selectedAlerts)
-            const newAlerts = _.differenceWith(selectedAlerts, alertDateRanges, _.isEqual)
-            setAlertDateRanges([...alertDateRanges, ...newAlerts])
-            for (const alertDateRange of newAlerts) {
-                const { data, error } = await supabase
-                    .from('alerts')
-                    .insert({
-                        start_date: alertDateRange.startDate,
-                        end_date: alertDateRange.endDate,
-                        river: alertDateRange.river,
-                        user_id: currentUser?.id,
-                    })
-                if (error) {
-                    console.error(error);
+            const [startDate, endDate] = dateRange;
+            if (startDate !== null && endDate !== null) {
+                const selectedAlerts = selectedRivers.flatMap((r) => ({
+                    startDate: startDate.toISOString().split('T')[0],
+                    endDate: endDate.toISOString().split('T')[0],
+                    river: r
+                }));
+                const newAlerts = _.differenceWith(selectedAlerts, alertDateRanges, _.isEqual)
+                if (alertDateRanges !== null) {
+                    setAlertDateRanges([...alertDateRanges, ...newAlerts])
+                }
+                for (const alertDateRange of newAlerts) {
+                    await supabase?.from('alerts')
+                        .insert({
+                            start_date: alertDateRange.startDate,
+                            end_date: alertDateRange.endDate,
+                            river: alertDateRange.river,
+                            user_id: currentUser?.id,
+                        })
                 }
             }
         } else {
@@ -92,19 +77,17 @@ export default function Alerts() {
         }
     };
 
-    const removeRow = async (index) => {
-        const newAlertDateRanges = alertDateRanges.filter((_, i) => i !== index);
-        setAlertDateRanges(newAlertDateRanges)
+    const removeRow = async (index: number) => {
+        if (alertDateRanges !== null) {
+            const newAlertDateRanges = alertDateRanges.filter((_, i) => i !== index);
+            setAlertDateRanges(newAlertDateRanges)
 
-        const deleteAlert = alertDateRanges.filter((_, i) => i === index)[0]
-        const { data, error } = await supabase
-            .from('alerts')
-            .delete()
-            .eq('start_date', deleteAlert.startDate)
-            .eq('end_date', deleteAlert.endDate)
-            .eq('river', deleteAlert.river)
-        if (error) {
-            console.error(error);
+            const deleteAlert = alertDateRanges.filter((_, i) => i === index)[0]
+            await supabase?.from('alerts')
+                .delete()
+                .eq('start_date', deleteAlert.startDate)
+                .eq('end_date', deleteAlert.endDate)
+                .eq('river', deleteAlert.river)
         }
 
     };
@@ -112,53 +95,62 @@ export default function Alerts() {
 
     return (
         <Center>
-            <Container size={'sm'}>
-                <Title>Alerts</Title>
+            <Container size={'xs'}>
+                <Title>Alert Selection</Title>
+                <Space my={'md'} />
+                <Text>Use the following date picker and river selector to setup alerts for specific rivers.</Text>
+                <Space my={'md'} />
                 <Stack>
-                    <Text>
-                        Select dates on the calendar below to subscribe to alerts for.
-                    </Text>
-                    <Center>
-                        <DatePicker
-                            numberOfColumns={2}
-                            type="range"
-                            value={dateRange}
-                            onChange={setDateRange} />
-                    </Center>
                     <MultiSelect
                         value={selectedRivers}
                         onChange={setSelectedRivers}
                         data={riverList}
                         placeholder="Select river(s)"
                     />
+                    <DatePicker
+                        type="range"
+                        numberOfColumns={2}
+                        value={dateRange}
+                        onChange={setDateRange}
+                    />
 
-                    <Button onClick={() => addRow(dateRange)}>Add Alert</Button>
-                    <Divider />
-                    <Title order={3}>Selected Alerts</Title>
-                    <Text>For user {currentUser?.email}</Text>
-                    <Table>
-                        <thead>
-                            <tr>
-                                <th>River</th>
-                                <th>Alert Start Date</th>
-                                <th>Alert End Date</th>
-                                <th>Remove</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {alertDateRanges.map((row, index) => (
-                                <tr key={index}>
-                                    <td>{row.river}</td>
-                                    <td>{row.startDate}</td>
-                                    <td>{row.endDate}</td>
-                                    <td>
-                                        <ActionIcon onClick={() => removeRow(index)}>
-                                            <IconX />
-                                        </ActionIcon>
-                                    </td>
-                                </tr>
-                            ))}
-                        </tbody>
+                    <Button onClick={() => {
+                        addRow(dateRange)
+                    }}>Add Alert</Button>
+                    <Space my={'md'} />
+                    <Group>
+                        <Title order={3}>Selected Alerts</Title>
+                        <Code>{currentUser !== null ? currentUser?.email : null}</Code>
+                    </Group>
+
+                    <Table highlightOnHover >
+                        <Table.Thead>
+                            <Table.Tr>
+                                <Table.Th align='left'>River</Table.Th>
+                                <Table.Th align='left'>Alert Start Date</Table.Th>
+                                <Table.Th align='left'>Alert End Date</Table.Th>
+                                <Table.Th align='center'>Remove</Table.Th>
+                            </Table.Tr>
+                        </Table.Thead>
+                        {
+                            currentUser !== null ?
+                                <Table.Tbody>
+                                    {alertDateRanges?.map((row, index) => (
+                                        <Table.Tr key={index}>
+                                            <Table.Td align="left">{row.river}</Table.Td>
+                                            <Table.Td align="left">{row.startDate.toString()}</Table.Td>
+                                            <Table.Td align="left">{row.endDate.toString()}</Table.Td>
+                                            <Table.Td align="center">
+                                                <ActionIcon color="red" variant="outline" onClick={() => removeRow(index)}>
+                                                    <IconX />
+                                                </ActionIcon>
+                                            </Table.Td>
+                                        </Table.Tr>
+                                    ))}
+                                </Table.Tbody>
+                                :
+                                <Table.Caption>Please login to see alerts</Table.Caption>
+                        }
                     </Table>
                 </Stack>
             </Container>
